+++ tests/typedb/mocks/config.ts
@@ -1,4 +1,12 @@
 import type { BormConfig } from '../../../src/index';
+import dotenv from 'dotenv';
+dotenv.config();
+
+// eslint-disable-next-line turbo/no-undeclared-env-vars
+const typeDbUrl = process.env.TYPE_DB_URL as string;
+if (!typeDbUrl) {
+	throw new Error('TYPE_DB_URL is not defined');
+}
 
 export const config: BormConfig = {
 	server: {
@@ -9,7 +17,7 @@ export const config: BormConfig = {
 			id: 'default',
 			provider: 'typeDB',
 			dbName: 'test',
-			url: 'localhost:1729',
+			url: typeDbUrl,
 		},
 	],
 };