+++ tests/unit/mutations/replaces.ts
@@ -1,7 +1,6 @@
-import 'jest';
-
 import { deepSort } from '../../helpers/matchers';
 import { createTest } from '../../helpers/createTest';
+import { expect, it } from 'vitest';
 
 export const testReplaceMutation = createTest('Mutation: Replaces', (ctx) => {
 	it('r1[replace] replace single roles in relation', async () => {