+++ tests/unit/queries/query.ts
@@ -1,4 +1,3 @@
-import 'jest';
 import { v4 as uuidv4 } from 'uuid';
 
 import { deepRemoveMetaData, deepSort, expectArraysInObjectToContainSameElements } from '../../helpers/matchers';
@@ -7,6 +6,7 @@ import type { TypeGen } from '../../../src/types/typeGen';
 import type { WithBormMetadata } from '../../../src/index';
 import type { UserType } from '../../types/testTypes';
 import { createTest } from '../../helpers/createTest';
+import { expect, it } from 'vitest';
 
 export const testQuery = createTest('Query', (ctx) => {
 	it('v1[validation] - $entity missing', async () => {