+++ changelog.md
@@ -2,6 +2,16 @@
 
 📝 following beta format X.Y.Z where Y = breaking change and Z = feature and fix. Later => FAIL.FEATURE.FIX
 
+## 0.10.3(2024-04-22)
+
+- Refacto: Pre queries dependencies refacto
+- Test: Removed jest dependencies
+
+## 0.10.2(2024-04-18)
+
+- Test: Migrated to vitest
+- Test: TypeDB url moved as env var
+
 ## 0.10.1(2024-04-17)
 
 - Feat: Transformations for mutations with $fields included.